Доработка сервисной логики

+++ portal/src/main/java/com/apteka/portal/services/GroupTaskService.java
@@ -6,15 +6,15 @@ import java.util.Objects;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.PathVariable;
 
 import com.apteka.portal.exceptions.DublicateGroupTaskException;
 import com.apteka.portal.exceptions.GroupTaskNotFoundException;
 import com.apteka.portal.exceptions.InvalidGroupTaskException;
-import com.apteka.portal.models.Client;
-import com.apteka.portal.models.ClientRole;
+import com.apteka.portal.models.AppUserDetails;
 import com.apteka.portal.models.GroupTask;
-import com.apteka.portal.models.UsersInApp;
+import com.apteka.portal.models.UserGroup;
+import com.apteka.portal.models.UserRole;
+import com.apteka.portal.models.UserType;
 import com.apteka.portal.repository.GroupTaskInterface;
 import lombok.RequiredArgsConstructor;
 
@@ -32,60 +32,71 @@ public class GroupTaskService {
     }
 
     @Transactional(readOnly = true)
-    public GroupTask getOne(@PathVariable Integer id) {
+    public GroupTask getOne(Integer id) {
         return groupTaskInterface.findById(id)
                 .orElseThrow(() -> new GroupTaskNotFoundException(id));
     }
 
     @Transactional
-    public GroupTask create(String name, UsersInApp currentUser) {
-        validateGroupTaskName(name);
-        validateCanCreateGroupTask(currentUser);
+    public GroupTask create(String name, Integer userGroupId) {
+
+        AppUserDetails currentUser = SecurityUtils.getCurrentUser();
+        UserGroup userGroup = userGroupService.getOne(userGroupId);
+
+        validateGroupTaskName(name, userGroup);
+        validateBossOrAdmin(currentUser, userGroup);
 
         return groupTaskInterface.save(GroupTask.builder()
                 .name(name.strip())
-                .userGroup(currentUser.getUserGroup())
+                .userGroup(Objects.requireNonNull(userGroup))
                 .build());
     }
 
     @Transactional
-    public GroupTask update(Integer id, String name, UsersInApp currentUser) {
+    public GroupTask update(Integer id, String name) {
+
+        AppUserDetails currentUser = SecurityUtils.getCurrentUser();
         GroupTask upGroup = getOne(id);
 
-        validateCanUpdateGroupTask(upGroup, currentUser);
-        validateGroupTaskName(name);
+        validateBossOrAdmin(currentUser, upGroup.getUserGroup());
+
+        if (Objects.equals(name, upGroup.getName())) {
+            return upGroup;
+        }
+
+        validateGroupTaskName(name, upGroup.getUserGroup());
 
         upGroup.setName(name.strip());
         return groupTaskInterface.save(upGroup);
     }
 
     @Transactional
-    public void delete(Integer id, UsersInApp currentUser) {
+    public void delete(Integer id) {
+        AppUserDetails currentUser = SecurityUtils.getCurrentUser();
         GroupTask deletedTask = getOne(id);
-        validateCanUpdateGroupTask(deletedTask, currentUser);
+        validateBossOrAdmin(currentUser, deletedTask.getUserGroup());
         groupTaskInterface.deleteById(id);
     }
 
-    private void validateGroupTaskName(String name) {
+    private void validateGroupTaskName(String name, UserGroup userGroup) {
         if (name == null || name.isBlank())
             throw new InvalidGroupTaskException(name);
-        if (groupTaskInterface.findByName(name).isPresent())
+        if (groupTaskInterface.findByNameAndUserGroupId(name, userGroup.getId()).isPresent())
             throw new DublicateGroupTaskException(name);
     }
 
-    private void validateCanUpdateGroupTask(GroupTask groupTask, UsersInApp currentUser) {
-        if(currentUser instanceof Client client) {
-            if (client.getRole() == ClientRole.ADMIN) return;
-            if (Objects.equals(client.getRole(), ClientRole.BOSS) && Objects.equals(client.getUserGroup().getId(), groupTask.getUserGroup().getId())) return;
-            throw new AccessDeniedException("Изменить списи работ могут только начальники своего отдела");
+    private void validateBossOrAdmin(AppUserDetails currentUser, UserGroup userGroup) {
+        if (currentUser.getType() != UserType.CLIENT) {
+            throw new AccessDeniedException("Недопустимый тип пользователя");
         }
-    }
 
-    private void validateCanCreateGroupTask(UsersInApp currentUser) {
-        if (currentUser instanceof Client client) {
-            if (client.getRole() == ClientRole.ADMIN) return;
-            if(Objects.equals(client.getRole(), ClientRole.BOSS)) return;
-            throw new AccessDeniedException("Созда списи работ могут только начальники своего отдела");
+        boolean isAdmin = currentUser.getRoles().contains(UserRole.ADMIN);
+        boolean isBossOfGroup = currentUser.getRoles().contains(UserRole.BOSS)
+                && Objects.equals(currentUser.getUserGroup().getId(), userGroup.getId());
+
+        if (isAdmin || isBossOfGroup) {
+            return;
         }
+        throw new AccessDeniedException("Создать списки работ могут только начальники своего отдела");
     }
 }
